pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "control-habits"
version = "0.1.0"
description = "Сервис трекинга дел и привычек: веб + Telegram-бот"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.0",
    "apscheduler>=3.10",
    "fastapi>=0.100",
    "httpx>=0.25",
    "pydantic-settings>=2.0",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.13",
    "psycopg[binary]>=3",
    "uvicorn[standard]>=0.22",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
